url_shortener/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic-settings